tweaks: adjust the type of two arguments
Convert digits() to take a ssize_t instead of an int, since it's
used on ssize_t line numbers.  And properly use the long modifier
when displaying a line number.

Also, conditionalize the digits() prototype.

+++ src/utils.c
@@ -54,7 +54,7 @@ void get_homedir(void)
 
 #ifdef ENABLE_LINENUMBERS
 /* Return the number of digits that the given integer n takes up. */
-int digits(int n)
+int digits(ssize_t n)
 {
     if (n < 100000) {
         if (n < 1000) {
